Add files via upload
EKG header and cpp, simply reads from the sensor and outputs the values. Need to find out at what rate it must measure to ensure that the peaks are found so that a timer between two peaks actually measures the time between those two peaks and not more peaks, this is to find the pulse.