Add files via upload
This is a macro used to debug. It will print out the file and line where the macro function is placed. It will be used regularly to see what parts of the code that runs, and when I am finished with debugging it will simply not be "a part of the code" because I will remove the #define DEBUGGER line in the print_debug().h